All Active Messages RecyclerView Implementation
Created the recycler view for all active messages and attached it to MessagesActivity. For the recycler view, I created a MessageItem data class for what is displayed in the recycler view, as well as created an adapter for the recycler view. I also created an activity called ConversationActivity for when the user clicks into a conversation for a specific user.